please lint
git-svn-id: file:///svn/unbound/trunk@1829 be551aaa-1e26-0410-a405-d3ace91eadb9

+++ util/log.c
@@ -131,7 +131,7 @@ log_init(const char* filename, int use_syslog, const char* chrootdir)
 	}
 #ifndef UB_ON_WINDOWS
 	/* line buffering does not work on windows */
-	setvbuf(f, NULL, _IOLBF, 0);
+	setvbuf(f, NULL, (int)_IOLBF, 0);
 #endif
 	logfile = f;
 }
